refactor: `TaskItem`の切り出し
